pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "coral-v4"
version = "4.0.0"
description = "CORAL v4: COrtical Reasoning via Abstraction Layers"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "pydantic>=2.0.0",
    "einops>=0.7.0",
    "tqdm>=4.65.0",
    "huggingface_hub>=0.20.0",
    "argdantic>=0.5.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["coral*"]
